# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SteaMates API",
    "description": "API REST de SteaMates — plataforma social para jugadores de Steam. Permite consultar perfiles, bibliotecas, estadísticas, precios de juegos, listas de la comunidad, sesiones de juego y moderación.",
    "version": "1.0.0"
  },
  "host": "backend-5rwo.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT obtenido al iniciar sesión con Steam. Formato: \"Bearer <token>\""
    }
  },
...
  "paths": {
...
    "/api/chat/market-recommendations": {
      "post": {
        "tags": ["Chat"],
        "summary": "Recomendaciones de mercado con IA",
        "description": "Genera recomendaciones de juegos en oferta personalizadas usando la biblioteca de Steam del usuario y la IA de Groq.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "steamId": { "type": "string", "description": "SteamID64 del usuario" },
                "limit":   { "type": "integer", "description": "Número máximo de recomendaciones (default 5)" }
              },
              "required": ["steamId"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Lista de recomendaciones generadas" },
          "400": { "description": "steamId no proporcionado" },
          "503": { "description": "API de Groq no configurada" }
        }
      }
    },
    "/api/chat/message": {
      "post": {
        "tags": ["Chat"],
        "summary": "Enviar mensaje al chatbot",
        "description": "Envía un mensaje al asistente de IA. Soporta contexto de Steam, historial de sesión e imágenes. Límite por usuario: 3 consultas/minuto (1ª con Llama 3.3 70B Versatile, 2ª y 3ª con Llama 4 Scout 17B). Las imágenes siempre usan Llama 4 Scout 17B.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "message":             { "type": "string", "description": "Texto del mensaje" },
                "sessionId":           { "type": "string", "description": "ID de sesión de chat para mantener historial" },
                "userId":              { "type": "string", "description": "ID del usuario" },
                "steamId":             { "type": "string", "description": "SteamID64 para incluir contexto de Steam" },
                "screenContext":       { "type": "string", "description": "Contexto adicional de la pantalla actual" },
                "includeSteamContext": { "type": "boolean", "description": "Si incluir datos de Steam en el contexto" },
                "image":               { "type": "string", "description": "Imagen en base64 para análisis visual" }
              },
              "required": ["message"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Respuesta del asistente" },
          "400": { "description": "Mensaje vacío o inválido" },
          "401": { "description": "No autorizado" },
          "429": { "description": "Límite de usuario superado (3 consultas/min). La respuesta incluye 'error' con la hora exacta de fin de cooldown y 'retryAfter' en segundos" },
          "500": { "description": "Error interno" },
          "503": { "description": "API de Groq no disponible" }
        }
      }
    },
    "/api/chat/history/{sessionId}": {
      "get": {
        "tags": ["Chat"],
        "summary": "Historial de chat",
        "description": "Devuelve el historial de mensajes de una sesión de chat.",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "type": "string", "description": "ID de la sesión de chat" }
        ],
        "responses": {
          "200": { "description": "Historial de mensajes" },
          "404": { "description": "Sesión no encontrada" },
          "500": { "description": "Error interno" }
        }
      }
    },
...
  }
}